+++ lib/common/templates/templates.dart
@@ -102,4 +102,108 @@ class AppPages {
   ];
 }
 ''';
+
+  static String main(String importPrefix) {
+    return '''
+import 'package:${importPrefix}/app.dart';
+import 'package:flutter/material.dart';
+import 'package:flutter/services.dart';
+// import 'package:get/get.dart';
+
+Future<void> main() async {
+  WidgetsFlutterBinding.ensureInitialized();
+
+  // Initialize Services
+  // await Get.putAsync(() => StorageService().init());
+  _hideSystemUI();
+  runApp(const App());
+}
+
+void _hideSystemUI() {
+  SystemChrome.setSystemUIOverlayStyle(
+    const SystemUiOverlayStyle(
+      statusBarColor: Colors.transparent,
+      statusBarBrightness: Brightness.light,
+      systemStatusBarContrastEnforced: true,
+    ),
+  );
+
+  SystemChrome.setPreferredOrientations(<DeviceOrientation>[
+    DeviceOrientation.portraitUp,
+    DeviceOrientation.portraitDown,
+  ]);
+
+  SystemChrome.setSystemUIChangeCallback((systemOverlaysAreVisible) async {
+    if (systemOverlaysAreVisible) {
+      await Future.delayed(const Duration(seconds: 3));
+      SystemChrome.restoreSystemUIOverlays();
+    }
+  });
+  SystemChrome.setEnabledSystemUIMode(SystemUiMode.manual, overlays: []);
+}
+''';
+  }
+
+  static String app(String importPrefix) {
+    return '''
+import 'package:flutter/material.dart';
+import 'package:flutter_localizations/flutter_localizations.dart';
+import 'package:get/get.dart';
+import 'package:${importPrefix}/routes/app_pages.dart';
+
+class App extends StatefulWidget {
+  const App({super.key});
+
+  @override
+  State<App> createState() => AppState();
+}
+
+class AppState extends State<App> with WidgetsBindingObserver {
+  // final languageCtrl = Get.put(LanguageController());
+
+  @override
+  void initState() {
+    WidgetsBinding.instance.addObserver(this);
+    super.initState();
+  }
+
+  @override
+  void dispose() {
+    WidgetsBinding.instance.removeObserver(this);
+    super.dispose();
+  }
+
+  @override
+  void didChangeAppLifecycleState(AppLifecycleState state) {
+    if (state == AppLifecycleState.resumed) {
+      // if (_state == AppLifecycleState.paused) {}
+    } else if (state == AppLifecycleState.paused) {}
+    super.didChangeAppLifecycleState(state);
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return GetMaterialApp(
+      initialRoute: AppPages.INITIAL, // Ensure this maps to a valid route
+      getPages: AppPages.routes,
+      navigatorKey: Get.key,
+      // locale: Locale(languageCtrl.currentLanguage.value),
+      // translations: LanguageUtil(),
+      title: "App Name", // GlobalConstants.kAppName
+      // supportedLocales: GlobalConstants.supportedLocales,
+      theme: ThemeData.light(), // MTheme.light
+      darkTheme: ThemeData.dark(), // MTheme.dark
+      // themeMode: StorageService.to.isDarkMode ? ThemeMode.dark : ThemeMode.light,
+      debugShowCheckedModeBanner: false,
+      localizationsDelegates: const [
+        GlobalMaterialLocalizations.delegate,
+        GlobalWidgetsLocalizations.delegate,
+        GlobalCupertinoLocalizations.delegate,
+      ],
+      // fallbackLocale: LanguageUtil.fallbackLocale,
+    );
+  }
+}
+''';
+  }
 }